refactor(connections): introduce function to dry getting model and label
